feat: add build.sh clean option

+++ build.sh
@@ -13,25 +13,23 @@ function runqmk() {
     $*
 }
 
-# 0. Clean
-rm -rf .build
+if [[ $1 == 'clean' ]]; then
+    # 0. Clean
+    runqmk make $KEYBOARD:$KEYMAP:clean
+    rm -rf .build
+    git clean -xfd
+    git submodule foreach --recursive git clean -xfd
+    git checkout -- .
+else
+    # 1. Copy my files into their appropriate locations inside QMK
+    mkdir -p $QMK_DIR/keyboards/$KEYBOARD/keymaps/$KEYMAP
+    cp -r $KEYMAP $QMK_DIR/keyboards/$KEYBOARD/keymaps/
 
-# 1. Copy my files into their appropriate locations inside QMK
-mkdir -p $QMK_DIR/keyboards/$KEYBOARD/keymaps/$KEYMAP
-cp -r $KEYMAP $QMK_DIR/keyboards/$KEYBOARD/keymaps/
+    # 2. Make teensy build
+    cd $QMK_DIR
+    runqmk make $KEYBOARD:$KEYMAP
 
-# 2. Run string replacement to tweak EZ defaults (this is an alternative to
-# copying the overrides dir)
-
-# 3. Make clean
-cd $QMK_DIR
-runqmk make $KEYBOARD:$KEYMAP:clean
-
-# 4. Make teensy build
-runqmk make $KEYBOARD:$KEYMAP
-
-# 5. Clean up
-cp -r .build ../.build
-#git clean -f -d
-#git checkout -- .
+    # 3. make it accessible
+    cp -r .build ../.build
+fi
